+++ unlock-app/src/components/interface/checkout/CardConfirmationCheckout.tsx
@@ -55,14 +55,20 @@ export const CardConfirmationCheckout = ({
     setIsAdvanced,
     onRecipientChange,
     advancedRecipientValid,
-    recipient: customAddress,
+    recipient,
     checkingRecipient,
   } = useAdvancedCheckout()
 
-  const totalPrice: number = Object.values(
-    lock.fiatPricing.usd as number
-  ).reduce((s: number, x: number): number => s + x, 0) as number
-  const fee = totalPrice - lock.fiatPricing.usd.keyPrice
+  let totalPrice: number = 0
+  let fee: number = 0
+  if (lock.fiatPricing?.usd) {
+    totalPrice = Object.values(lock.fiatPricing.usd as number).reduce(
+      (s: number, x: number): number => s + x,
+      0
+    ) as number
+    fee = totalPrice - lock.fiatPricing.usd.keyPrice
+  }
+
   const formattedPrice = (totalPrice / 100).toFixed(2)
 
   useEffect(() => {
@@ -97,7 +103,7 @@ export const CardConfirmationCheckout = ({
         lock.address,
         network,
         formattedPrice,
-        customAddress
+        recipient || account
       )
       if (hash) {
         emitTransactionInfo({