+++ packages/unlock-js/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changes
 
+# Next patch
+
+- Non expired keys are represented as -1
+- Gas price calculation fixes
+- Adding support for unlimited durations when granting keys
+
 # 0.22.2
 
 - Add support for setExpirationDuration